Batch the query and delete by primary key in EntityUsageTable::pruneStaleUsages
Bug: T114654

+++ client/includes/Usage/Sql/EntityUsageTable.php
@@ -319,15 +319,31 @@ class EntityUsageTable {
 
 		$old = $this->queryUsages( $pageId, '<', $lastUpdatedBefore );
 
-		// XXX: we may want to batch this, based on the data in $old
-		$this->connection->delete(
-			$this->tableName,
-			array(
-				'eu_page_id' => (int)$pageId,
-				'eu_touched < ' . $this->connection->addQuotes( $lastUpdatedBefore ),
-			),
-			__METHOD__
-		);
+		do {
+			// @TODO: This is very similar to what getAffectedRowIds does.
+			$res = $this->connection->selectFieldValues(
+				$this->tableName,
+				'eu_row_id',
+				array(
+					'eu_page_id' => (int)$pageId,
+					'eu_touched < ' . $this->connection->addQuotes( $lastUpdatedBefore ),
+				),
+				__METHOD__,
+				array( 'LIMIT' => $this->batchSize )
+			);
+
+			if ( !$res ) {
+				break;
+			}
+
+			$this->connection->delete(
+				$this->tableName,
+				array(
+					'eu_row_id' => $res,
+				),
+				__METHOD__
+			);
+		} while ( count( $res ) === $this->batchSize );
 
 		return $old;
 	}
